Resolve a promise once instead of for each case
Currently, the promise is resolved multiple times, once for every file written. it should ideally resolve like the run benchmark case - only once for all the cases.

In this commit, the promise is rejected on the first error encountered - an exception is thrown, and any other promise catchers see that as a promise rejection.